[example][ff-pm] Split pmproblem and properties for 3D case
