+++ examples/ff-pm/flow-over-cube-3d/pmproblem-reversed.hh
@@ -24,65 +24,14 @@
 #ifndef DUMUX_DARCY_SUBPROBLEM_HH
 #define DUMUX_DARCY_SUBPROBLEM_HH
 
-#include <dune/grid/yaspgrid.hh>
-
 #include <dumux/common/numeqvector.hh>
 
-#include <dumux/discretization/cctpfa.hh>
-
-#include <dumux/porousmediumflow/1p/model.hh>
 #include <dumux/porousmediumflow/problem.hh>
-//#include "../../common/porousmediumflowproblemwithgravity.hh"
-
-#include "1pspatialparams.hh"
-
-#include <dumux/material/components/simpleh2o.hh>
-#include <dumux/material/fluidsystems/1pliquid.hh>
 
 #include <dumux-precice/couplingadapter.hh>
 
 namespace Dumux
 {
-template<class TypeTag>
-class DarcySubProblem;
-
-namespace Properties
-{
-// Create new type tags
-namespace TTag
-{
-struct DarcyOneP {
-    using InheritsFrom = std::tuple<OneP, CCTpfaModel>;
-};
-}  // end namespace TTag
-
-// Set the problem property
-template<class TypeTag>
-struct Problem<TypeTag, TTag::DarcyOneP> {
-    using type = Dumux::DarcySubProblem<TypeTag>;
-};
-
-// the fluid system
-template<class TypeTag>
-struct FluidSystem<TypeTag, TTag::DarcyOneP> {
-    using Scalar = GetPropType<TypeTag, Properties::Scalar>;
-    using type =
-        FluidSystems::OnePLiquid<Scalar, Dumux::Components::SimpleH2O<Scalar>>;
-};
-
-// Set the grid type
-template<class TypeTag>
-struct Grid<TypeTag, TTag::DarcyOneP> {
-    using type = Dune::YaspGrid<3>;
-};
-
-template<class TypeTag>
-struct SpatialParams<TypeTag, TTag::DarcyOneP> {
-    using type = OnePSpatialParams<GetPropType<TypeTag, GridGeometry>,
-                                   GetPropType<TypeTag, Scalar>>;
-};
-
-}  // end namespace Properties
 
 /*!
  * \brief The porous medium flow sub problem
